requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
amqp==1.4.9               # via -r requirements/base.txt, kombu
anyjson==0.3.3            # via -r requirements/base.txt, kombu
astroid==2.3.3            # via pylint, pylint-celery
attrs==19.3.0             # via pytest
billiard==3.3.0.23        # via -r requirements/base.txt, celery
celery==3.1.26.post2      # via -r requirements/base.txt
certifi==2020.4.5.1       # via -r requirements/base.txt, requests
chardet==3.0.4            # via -r requirements/base.txt, requests
click-log==0.3.2          # via edx-lint
click==7.1.2              # via click-log, code-annotations, edx-lint
code-annotations==0.3.4   # via -r requirements/test.in
coreapi==2.3.3            # via -r requirements/base.txt, django-rest-swagger, openapi-codec
coreschema==0.0.4         # via -r requirements/base.txt, coreapi
coverage==5.1             # via -r requirements/test.in, pytest-cov
defusedxml==0.6.0         # via -r requirements/base.txt, python3-openid, social-auth-core
django-cors-headers==3.3.0  # via -r requirements/base.txt
django-dynamic-fixture==3.1.0  # via -r requirements/test.in
django-extensions==2.2.9  # via -r requirements/base.txt
django-model-utils==4.0.0  # via -r requirements/base.txt
django-rest-swagger==2.2.0  # via -r requirements/base.txt
django-simple-history==2.10.0  # via -r requirements/base.txt
django-waffle==0.20.0     # via -r requirements/base.txt, edx-django-utils, edx-drf-extensions
djangorestframework==3.11.0  # via -r requirements/base.txt, django-rest-swagger, drf-jwt, drf-nested-routers, edx-drf-extensions, rest-condition
drf-jwt==1.14.0           # via -r requirements/base.txt, edx-drf-extensions
drf-nested-routers==0.91  # via -r requirements/base.txt
edx-auth-backends==3.1.0  # via -r requirements/base.txt
edx-django-utils==3.2.2   # via -r requirements/base.txt, edx-drf-extensions
edx-drf-extensions==6.0.0  # via -r requirements/base.txt
edx-lint==1.4.1           # via -r requirements/test.in
edx-opaque-keys==2.1.0    # via -r requirements/base.txt, edx-drf-extensions
edx-rest-api-client==1.9.2  # via -r requirements/base.txt
factory-boy==2.12.0       # via -r requirements/test.in
faker==4.1.0              # via factory-boy
future==0.18.2            # via -r requirements/base.txt, pyjwkest
idna==2.9                 # via -r requirements/base.txt, requests
importlib-metadata==1.6.0  # via pluggy, pytest
isort==4.3.21             # via pylint
itypes==1.2.0             # via -r requirements/base.txt, coreapi
jinja2==2.11.2            # via -r requirements/base.txt, code-annotations, coreschema
kombu==3.0.37             # via -r requirements/base.txt, celery
lazy-object-proxy==1.4.3  # via astroid
markupsafe==1.1.1         # via -r requirements/base.txt, jinja2
mccabe==0.6.1             # via pylint
mock==3.0.5               # via -r requirements/test.in
more-itertools==8.3.0     # via pytest
newrelic==5.12.1.141      # via -r requirements/base.txt, edx-django-utils
oauthlib==3.1.0           # via -r requirements/base.txt, requests-oauthlib, social-auth-core
openapi-codec==1.3.2      # via -r requirements/base.txt, django-rest-swagger
packaging==20.4           # via pytest
pbr==5.4.5                # via -r requirements/base.txt, stevedore
pluggy==0.13.1            # via pytest
psutil==1.2.1             # via -r requirements/base.txt, edx-django-utils
py==1.8.1                 # via pytest
pycryptodomex==3.9.7      # via -r requirements/base.txt, pyjwkest
pyjwkest==1.4.2           # via -r requirements/base.txt, edx-drf-extensions
pyjwt==1.7.1              # via -r requirements/base.txt, drf-jwt, edx-auth-backends, edx-rest-api-client, social-auth-core
pylint-celery==0.3        # via edx-lint
pylint-django==2.0.11     # via edx-lint
pylint-plugin-utils==0.6  # via pylint-celery, pylint-django
pylint==2.4.2             # via edx-lint, pylint-celery, pylint-django, pylint-plugin-utils
pymongo==3.10.1           # via -r requirements/base.txt, edx-opaque-keys
pyparsing==2.4.7          # via packaging
pytest-cov==2.9.0         # via -r requirements/test.in
pytest-django==3.9.0      # via -r requirements/test.in
pytest==5.4.2             # via pytest-cov, pytest-django
python-dateutil==2.8.1    # via -r requirements/base.txt, edx-drf-extensions, faker
python-slugify==4.0.0     # via code-annotations
python3-openid==3.1.0     # via -r requirements/base.txt, social-auth-core
pytz==2020.1              # via -r requirements/base.txt, celery, django
pyyaml==5.3.1             # via code-annotations
redis==2.8                # via -r requirements/base.txt
requests-oauthlib==1.3.0  # via -r requirements/base.txt, social-auth-core
requests==2.23.0          # via -r requirements/base.txt, coreapi, edx-drf-extensions, edx-rest-api-client, pyjwkest, requests-oauthlib, slumber, social-auth-core
rest-condition==1.0.3     # via -r requirements/base.txt, edx-drf-extensions
semantic-version==2.8.5   # via -r requirements/base.txt, edx-drf-extensions
simplejson==3.17.0        # via -r requirements/base.txt, django-rest-swagger
six==1.15.0               # via -r requirements/base.txt, astroid, django-dynamic-fixture, django-extensions, django-simple-history, django-waffle, edx-auth-backends, edx-drf-extensions, edx-lint, edx-opaque-keys, mock, packaging, pyjwkest, python-dateutil, social-auth-app-django, social-auth-core, stevedore
slumber==0.7.1            # via -r requirements/base.txt, edx-rest-api-client
social-auth-app-django==3.1.0  # via -r requirements/base.txt, edx-auth-backends
social-auth-core==3.2.0   # via -r requirements/base.txt, edx-auth-backends, social-auth-app-django
sqlparse==0.3.1           # via -r requirements/base.txt, django
stevedore==1.32.0         # via -r requirements/base.txt, code-annotations, edx-opaque-keys
text-unidecode==1.3       # via faker, python-slugify
typed-ast==1.4.1          # via astroid
uritemplate==3.0.1        # via -r requirements/base.txt, coreapi
urllib3==1.25.9           # via -r requirements/base.txt, requests
wcwidth==0.1.9            # via pytest
wrapt==1.11.2             # via astroid
zipp==1.2.0               # via -r requirements/base.txt, importlib-metadata
